Improve C# version, reduce register pressure

+++ csharp/PacketProcessors.cs
@@ -3,7 +3,7 @@ using TinyNF.Unsafe;
 
 namespace TinyNF
 {
-    public delegate void PacketProcessor(ref PacketData data, uint length, Array256<uint> outputs);
+    public delegate void PacketProcessor(ref PacketData data, ushort length, Array256<ushort> outputs);
 
-    public delegate void SafePacketProcessor(ref PacketData data, uint length, Span<uint> outputs);
+    public delegate void SafePacketProcessor(ref PacketData data, ushort length, Span<ushort> outputs);
 }
